[18] Admin role (#12)

+++ src/main/java/com/ensiasit/projectx/services/ContestServiceImpl.java
@@ -2,12 +2,17 @@ package com.ensiasit.projectx.services;
 
 import com.ensiasit.projectx.dto.ContestDto;
 import com.ensiasit.projectx.dto.UserContestRoleDto;
+import com.ensiasit.projectx.exceptions.ForbiddenException;
 import com.ensiasit.projectx.exceptions.NotFoundException;
+import com.ensiasit.projectx.mappers.ContestMapper;
 import com.ensiasit.projectx.models.Contest;
+import com.ensiasit.projectx.models.UserContestRole;
 import com.ensiasit.projectx.repositories.ContestRepository;
 import com.ensiasit.projectx.repositories.UserContestRoleRepository;
+import com.ensiasit.projectx.utils.RoleEnum;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -17,88 +22,69 @@ import java.util.Optional;
 public class ContestServiceImpl implements ContestService {
     private final ContestRepository contestRepository;
     private final UserContestRoleRepository userContestRoleRepository;
+    private final AdminService adminService;
+    private final ContestMapper contestMapper;
 
     @Override
     public List<ContestDto> getAll() {
         return contestRepository.findAll().stream()
-                .map(contest -> ContestDto.builder()
-                        .id(contest.getId())
-                        .name(contest.getName())
-                        .startTime(contest.getStartTime())
-                        .endTime(contest.getEndTime())
-                        .freezeTime(contest.getFreezeTime())
-                        .unfreezeTime(contest.getUnfreezeTime())
-                        .publicScoreboard(contest.isPublicScoreboard())
-                        .build())
+                .map(contestMapper::toContestDto)
                 .toList();
     }
 
     @Override
-    public ContestDto createContest(ContestDto contestDto) {
-        Contest contest = contestRepository.save(Contest.builder()
-                .name(contestDto.getName())
-                .startTime(contestDto.getStartTime())
-                .endTime(contestDto.getEndTime())
-                .freezeTime(contestDto.getFreezeTime())
-                .unfreezeTime(contestDto.getUnfreezeTime())
-                .publicScoreboard(contestDto.isPublicScoreboard())
-                .build());
+    public ContestDto createContest(String userEmail, ContestDto contestDto) {
+        if (!adminService.isAdmin(userEmail)) {
+            throw new ForbiddenException("User has no write access.");
+        }
 
-        return ContestDto.builder()
-                .id(contest.getId())
-                .name(contest.getName())
-                .startTime(contest.getStartTime())
-                .endTime(contest.getEndTime())
-                .freezeTime(contest.getFreezeTime())
-                .unfreezeTime(contest.getUnfreezeTime())
-                .publicScoreboard(contest.isPublicScoreboard())
-                .build();
+        Contest contest = contestRepository.save(contestMapper.fromContestDto(contestDto));
+
+        return contestMapper.toContestDto(contest);
     }
 
     @Override
     public ContestDto getContest(long id) {
         Contest contest = extractContest(id);
 
-        return ContestDto.builder()
-                .id(contest.getId())
-                .name(contest.getName())
-                .startTime(contest.getStartTime())
-                .endTime(contest.getEndTime())
-                .freezeTime(contest.getFreezeTime())
-                .unfreezeTime(contest.getUnfreezeTime())
-                .publicScoreboard(contest.isPublicScoreboard())
-                .build();
+        return contestMapper.toContestDto(contest);
     }
 
+    @Transactional
     @Override
-    public ContestDto deleteContest(long id) {
+    public ContestDto deleteContest(String userEmail, long id) {
         Contest contest = extractContest(id);
+
+        if (userHasNoWriteAccess(userEmail, id)) {
+            throw new ForbiddenException("User has no write access.");
+        }
+
+        userContestRoleRepository.deleteAllByContestId(id);
         contestRepository.deleteById(id);
 
-        return ContestDto.builder()
-                .id(contest.getId())
-                .name(contest.getName())
-                .startTime(contest.getStartTime())
-                .endTime(contest.getEndTime())
-                .freezeTime(contest.getFreezeTime())
-                .unfreezeTime(contest.getUnfreezeTime())
-                .publicScoreboard(contest.isPublicScoreboard())
-                .build();
+        return contestMapper.toContestDto(contest);
     }
 
     @Override
     public List<UserContestRoleDto> getAllUserContests(String userEmail) {
+        if (adminService.isAdmin(userEmail)) {
+            return getAll().stream()
+                    .map(contestDto -> contestMapper.toUserContestRoleDto(contestDto, RoleEnum.ROLE_ADMIN))
+                    .toList();
+        }
+
         return userContestRoleRepository.findAllByUserEmail(userEmail).stream()
-                .map(userContestRole -> UserContestRoleDto.builder()
-                        .contest(userContestRole.getContest())
-                        .role(userContestRole.getRole())
-                        .build())
+                .map(contestMapper::toUserContestRoleDto)
                 .toList();
     }
 
     @Override
-    public ContestDto updateContest(long id, ContestDto payload) {
+    public ContestDto updateContest(String userEmail, long id, ContestDto payload) {
         Contest contest = extractContest(id);
+
+        if (userHasNoWriteAccess(userEmail, id)) {
+            throw new ForbiddenException("User has no write access.");
+        }
 
         contest.setName(payload.getName());
         contest.setStartTime(payload.getStartTime());
@@ -127,5 +113,18 @@ public class ContestServiceImpl implements ContestService {
         }
 
         return contestOptional.get();
+    }
+
+    private boolean userHasNoWriteAccess(String userEmail, long contestId) {
+        if (adminService.isAdmin(userEmail)) {
+            return false;
+        }
+
+        List<UserContestRole> userContestRoles = userContestRoleRepository.findAllByUserEmail(userEmail);
+
+        return userContestRoles.stream()
+                .filter(userContestRole -> userContestRole.getContest().getId() == contestId)
+                .noneMatch(userContestRole ->
+                        userContestRole.getRole().equals(RoleEnum.ROLE_ADMIN) || userContestRole.getRole().equals(RoleEnum.ROLE_MODERATOR));
     }
 }
